fix: move custom block layout to its own file
