+++ packages/ts-newskit/src/slices/comment-bucket-1/index.tsx
@@ -6,7 +6,7 @@ import {
   CommentCard,
   CommentCardProps
 } from '../../components/slices/comment-card';
-import { CustomBlockLayout } from '../shared/layouts';
+import { CustomBlockLayout } from '../shared';
 export interface CommentStackProps {
   clickHandler: ClickHandlerType;
   data: CommentCardProps[];
