+++ packages/ts-newskit/src/slices/shared/index.ts
@@ -1,4 +1,6 @@
 export { CustomStackLayout, WrappedStackLayout } from './layouts';
+export { CustomBlockLayout } from './custom-layout';
+
 export {
   ArticleStackSmall,
   ArticleStackLarge,
